base: Allow RRO-ing the certified props for PixelPropsUtils
This allows updating the props using a new RRO apk
Make this a "special" singleton that only initiates after getting a context that could read the props.
This way we only have to read em once per boot instead of on every app launch.
I am very aware this is kinda nasty, but this whole thing is anyways, use of reflection, breaking app trust etc
This is the best solution I've yet to see for this stupid hack. Better than reading the props for every app launch.
You'd also be surprised to know some even read a FILE here (ikr???). Surely this is MUCH faster.

Also add a check in PPU for when the overlay is disabled.
That allows disabling ALL PPU related hacks on a platform that doesn't need em (ex. vanilla)
Or for stupid modules... Unfortunately.